test: extract helpers in frontend tests
uvu seems to get confused about the reporting when importing from other
test files
